Replace csv.Reader with recordsReader interface.
The standard csv.Reader is not flexible enough in some cases when it's
used as a data provider in FileParser.

For example, you can have a file in which you want to skip certain
lines. Sometimes Reader.Comment may be suitable if the lines you want
to skip all start with the same character. Though in other more complex
cases there is no way to configure csv.Reader to skip needed lines.

Intriduction of the recordsReader should help to close this gap: you'll
just need to write a custom wrapper around csv.Reader with custom logic
for skipping lines that need to be skipped.